Don't include tournament internal names in match IDs.
These are no longer going to be used the way I thought they would
(at least in Tiltyard), so these are no longer necessary (and are
unused in the library). Clients are now responsible for tournament
name uniqueness.

This was a performance bottleneck (due to the possibility of
tournament names having hyphens and how regexes work) and is still
a bottleneck, but is now significantly faster. I should consider
also converting the entire match ID string into a single object
at some appropriate time, so users can then just grab int values
off the object.

Also bump the version number.