Change ranking message of single-elim stages in final state.
This used to describe the top player at the end of the stage as
simply "in contention", when all other players had been
eliminated. This now describes them as "winner". Theoretically
that could be odd if single-elimination is used somewhere other
than at the end of a tournament, but that would be an odd use
case to begin with.

This requires some tweaks to the StableTournamentFormatsTest. To
make this easier in the future, also modifies that test's storage
format to be more stable (i.e. store sets in a sorted order).